+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `tabindex="-1" to non-selected tabs to prevent them from being focusable when navigating with the keyboard.
 - Remove `tabindex="-1"` attribute from the "Scroll to start" and "Scroll to end" buttons, to allow them to be focusable when navigating with the keyboard.
+- Add `aria-orientation` attribute to the tablist element to indicate the orientation of the tabs.
 
 ## v2.3.1 (2024-08-22)